// src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    secondary?: string;
    info?: string;
    success?: string;
    accent?: string;
    warning?: string;
    error?: string;
    lightprimary?: string;
    lightsecondary?: string;
    lightsuccess?: string;
    lighterror?: string;
    lightwarning?: string;
    darkprimary?: string;
    darksecondary?: string;
    darkText?: string;
    lightText?: string;
    borderLight?: string;
    inputBorder?: string;
    containerBg?: string;
    surface?: string;
    background?: string;
    'on-surface-variant'?: string;
    facebook?: string;
    twitter?: string;
    linkedin?: string;
    gray100?: string;
    primary200?: string;
    secondary200?: string;
    white?: string;
    primary300?: string;
    primary400?: string;
    primary500?: string;
    primary600?: string;
    primary700?: string;
    primary800: string;
    primary900: string;
    secondary400?: string;
    secondary600?: string;
    secondary800?: string;

    // --- Chart Palette Addition ---
    chartColor1?: string;
    chartColor2?: string;
    chartColor3?: string;
    chartColor4?: string;
    chartColor5?: string;
    chartColor6?: string;
    chartColor7?: string;
    chartColor8?: string;
    chartColor9?: string;
    chartColor10?: string;
    chartColor11?: string;
    chartColor12?: string;
    chartColor13?: string;
    chartColor14?: string;
    chartColor15?: string;
    chartColor16?: string;
    chartColor17?: string;
    chartColor18?: string;
    chartColor19?: string;
    chartColor20?: string;
  };
};